refactor: renamed: api/models/db.go -> api/db/db.go #6
