+++ api/controllers/user_controller.go
@@ -2,13 +2,14 @@ package controllers
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/t2469/labor-management-system.git/db"
 	"github.com/t2469/labor-management-system.git/models"
 	"net/http"
 )
 
 func GetUsers(c *gin.Context) {
 	var users []models.User
-	if err := models.DB.Find(&users).Error; err != nil {
+	if err := db.DB.Find(&users).Error; err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 	}
 	c.JSON(http.StatusOK, users)
@@ -17,7 +18,7 @@ func GetUsers(c *gin.Context) {
 func GetUser(c *gin.Context) {
 	id := c.Param("id")
 	var user models.User
-	if err := models.DB.Preload("Attendances").First(&user, id).Error; err != nil {
+	if err := db.DB.Preload("Attendances").First(&user, id).Error; err != nil {
 		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
 		return
 	}
@@ -32,7 +33,7 @@ func CreateUser(c *gin.Context) {
 		return
 	}
 
-	if err := models.DB.Create(&user).Error; err != nil {
+	if err := db.DB.Create(&user).Error; err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 	}
