+++ api/db/db.go
@@ -1,4 +1,4 @@
-package models
+package db
 
 import (
 	"fmt"
